refactor: move message into src/api folder #591
