+++ server/src/api/controllers/settings.ts
@@ -1,7 +1,7 @@
 import { Request, Response } from "express";
 import _ from "lodash";
 import { logger } from "@logger";
-import { FIELD_INVALID, FIELD_REQUIRED, NOT_FOUND } from "@message";
+import { FIELD_INVALID, FIELD_REQUIRED, NOT_FOUND } from "src/api/message";
 import moment from "moment";
 import {
   basicTimeZone,
