+++ server/src/api/controllers/users.ts
@@ -12,11 +12,7 @@ import {
 import { cipher, decipher } from "../../../utils/crypto";
 
 import { logger } from "@logger";
-import {
-  EMAIL_IN_USE,
-  FIELD_REQUIRED,
-  USER_NOT_FOUND,
-} from "../../../@message";
+import { EMAIL_IN_USE, FIELD_REQUIRED, USER_NOT_FOUND } from "../message";
 
 //_____________________________________________________________________________
