+++ server/src/passport/googleStrategy.ts
@@ -7,7 +7,7 @@ import {
   EMAIL_IN_USE,
   SNSID_IN_USE,
   USER_NOT_FOUND,
-} from "../../@message";
+} from "src/api/message";
 
 const getEmail = (profile: Profile): string | undefined => {
   if (profile.emails) {
