+++ server/src/passport/naverStrategy.ts
@@ -2,7 +2,7 @@ import { Request } from "express";
 import passport, { Profile } from "passport";
 import { Strategy as NaverStrategy } from "passport-naver";
 import { User } from "@models/User";
-import { CONNECTED_ALREADY, EMAIL_IN_USE, SNSID_IN_USE } from "../../@message";
+import { CONNECTED_ALREADY, EMAIL_IN_USE, SNSID_IN_USE } from "../api/message";
 
 const getEmail = (profile: Profile): string | undefined => {
   if (profile.emails && profile.emails.length > 0) {
